% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_arm_moments}
\alias{compute_arm_moments}
\title{Compute arm-specific moments on a subset of data}
\usage{
compute_arm_moments(obs_subset, arm)
}
\arguments{
\item{obs_subset}{Subset of observed data}

\item{arm}{Which arm (1 or 0)}
}
\value{
List of arm-specific moments
}
\description{
Compute arm-specific moments on a subset of data
}
\keyword{internal}
