% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dip_pp.R
\name{compute_dip_pp_estimate}
\alias{compute_dip_pp_estimate}
\title{Compute DiP++ estimate for a fold}
\usage{
compute_dip_pp_estimate(obs, unobs, treated_idx, control_idx, lambda)
}
\description{
Compute DiP++ estimate for a fold
}
\keyword{internal}
