% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_dip.R
\name{compute_dt_dip_estimate}
\alias{compute_dt_dip_estimate}
\title{Compute D-T DiP estimate for a fold}
\usage{
compute_dt_dip_estimate(obs, unobs, treated_idx, control_idx, lambda1, lambda0)
}
\description{
Compute D-T DiP estimate for a fold
}
\keyword{internal}
