% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt.R
\name{compute_dt_estimate}
\alias{compute_dt_estimate}
\title{Compute D-T estimate for a fold}
\usage{
compute_dt_estimate(obs, unobs, treated_idx, control_idx, lambda1, lambda0)
}
\description{
Compute D-T estimate for a fold
}
\keyword{internal}
