% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_moments}
\alias{compute_moments}
\title{Compute all moments needed for estimation}
\usage{
compute_moments(data)
}
\arguments{
\item{data}{An msd_data object}
}
\value{
A list of sample moments
}
\description{
Compute all moments needed for estimation
}
\keyword{internal}
