% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi.R
\name{compute_ppi_estimate}
\alias{compute_ppi_estimate}
\title{Compute PPI++ estimate for a fold}
\usage{
compute_ppi_estimate(obs, unobs, treated_idx, control_idx, lambda)
}
\description{
Compute PPI++ estimate for a fold
}
\keyword{internal}
