% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cov_pop}
\alias{cov_pop}
\title{Compute sample covariance (population formula, divide by n)}
\usage{
cov_pop(x, y)
}
\arguments{
\item{x}{A numeric vector}

\item{y}{A numeric vector}
}
\value{
Sample covariance using n denominator
}
\description{
Compute sample covariance (population formula, divide by n)
}
\keyword{internal}
