% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dip_pp.R
\name{estimate_lambda_dip_pp}
\alias{estimate_lambda_dip_pp}
\title{Estimate lambda for DiP++ (single lambda)}
\usage{
estimate_lambda_dip_pp(obs, treated_idx, control_idx, unobs, m)
}
\description{
Estimate lambda for DiP++ (single lambda)
}
\keyword{internal}
