% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt.R
\name{estimate_lambda_dt}
\alias{estimate_lambda_dt}
\title{Estimate arm-specific lambdas for D-T}
\usage{
estimate_lambda_dt(obs, treated_idx, control_idx, unobs, m1, m0)
}
\description{
Estimate arm-specific lambdas for D-T
}
\keyword{internal}
