% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_dip.R
\name{estimate_lambda_dt_dip}
\alias{estimate_lambda_dt_dip}
\title{Estimate arm-specific lambdas for D-T DiP}
\usage{
estimate_lambda_dt_dip(obs, treated_idx, control_idx, unobs, m)
}
\description{
Estimate arm-specific lambdas for D-T DiP
}
\keyword{internal}
