% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi.R
\name{estimate_lambda_ppi}
\alias{estimate_lambda_ppi}
\title{Estimate lambda for PPI++ (single lambda across arms)}
\usage{
estimate_lambda_ppi(obs, treated_idx, control_idx, unobs, m1, m0)
}
\description{
Estimate lambda for PPI++ (single lambda across arms)
}
\keyword{internal}
