% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_design.R
\name{expected_variance}
\alias{expected_variance}
\title{Calculate expected variance for a given design}
\usage{
expected_variance(estimator, n_obs, n_unobs, treatment_prob, moments)
}
\description{
Calculate expected variance for a given design
}
\keyword{internal}
