% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_design.R
\name{extract_pilot_moments}
\alias{extract_pilot_moments}
\title{Extract population moments from pilot data}
\usage{
extract_pilot_moments(pilot_data)
}
\description{
Extract population moments from pilot data
}
\keyword{internal}
