% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedsubjects-package.R
\docType{package}
\name{mixedsubjects-package}
\alias{mixedsubjects}
\alias{mixedsubjects-package}
\title{mixedsubjects: Causal Inference in Experiments with Mixed-Subjects Designs}
\description{
Implements seven estimators for average treatment effect (ATE) estimation in mixed-subjects designs (MSDs), where human subjects data is augmented with predictions from large language models (LLMs). Includes Difference-in-Means, GREG, PPI++, Doubly-Tuned, Difference-in-Predictions (DiP), DiP++, and D-T DiP estimators. Provides point estimates, variance estimation via delta-method or bootstrap, and optimal design selection for budget allocation between human observations and LLM predictions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://klintkanopka.com/mixedsubjects/}
  \item Report bugs at \url{https://github.com/klintkanopka/mixedsubjects/issues}
}

}
\author{
\strong{Maintainer}: Klint Kanopka \email{klint.kanopka@nyu.edu}

Authors:
\itemize{
  \item Austin van Loon \email{avanloon@mit.edu}
}

Other contributors:
\itemize{
  \item Yuan Huang \email{yh2741@nyu.edu} [contributor]
}

}
\keyword{internal}
