% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{new_msd_result}
\alias{new_msd_result}
\title{Utility Functions for Mixed-Subjects Design}
\usage{
new_msd_result(
  estimate,
  variance,
  se,
  ci_lower,
  ci_upper,
  method,
  lambda = NULL,
  n1 = NULL,
  n0 = NULL,
  m = NULL,
  conf_level = 0.95,
  additional = list()
)
}
\arguments{
\item{estimate}{Point estimate of the ATE}

\item{variance}{Estimated variance}

\item{se}{Standard error}

\item{ci_lower}{Lower bound of confidence interval}

\item{ci_upper}{Upper bound of confidence interval}

\item{method}{Name of the estimation method}

\item{lambda}{Tuning parameter(s), if applicable}

\item{n1}{Number of treated observed units}

\item{n0}{Number of control observed units}

\item{m}{Number of unobserved (unlabeled) units}

\item{conf_level}{Confidence level used}

\item{additional}{Additional method-specific information}
}
\value{
An S3 object of class "msd_result"
}
\description{
Internal helper functions for the mixedsubjects package.
}
\keyword{Create}
\keyword{an}
\keyword{internal}
\keyword{msd_result}
\keyword{object}
