% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{print.msd_data}
\alias{print.msd_data}
\title{Print method for msd_data}
\usage{
\method{print}{msd_data}(x, ...)
}
\arguments{
\item{x}{An msd_data object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns x
}
\description{
Print method for msd_data
}
