% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_design.R
\name{print.msd_design}
\alias{print.msd_design}
\title{Print method for msd_design}
\usage{
\method{print}{msd_design}(x, digits = 4, ...)
}
\arguments{
\item{x}{An msd_design object.}

\item{digits}{Number of digits to display.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for msd_design
}
