% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.msd_result}
\alias{print.msd_result}
\title{Print method for msd_result}
\usage{
\method{print}{msd_result}(x, digits = 4, ...)
}
\arguments{
\item{x}{An msd_result object}

\item{digits}{Number of digits to display}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns x
}
\description{
Print method for msd_result
}
