% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{process_combined_data}
\alias{process_combined_data}
\title{Process combined dataframe}
\usage{
process_combined_data(data, y_col, s0_col, s1_col, d_col)
}
\description{
Process combined dataframe
}
\keyword{internal}
