% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{process_separate_data}
\alias{process_separate_data}
\title{Process separate dataframes}
\usage{
process_separate_data(
  observed_df,
  unobserved_df,
  y_col,
  d_col_obs,
  s0_col_obs,
  s1_col_obs,
  d_col_unobs,
  s0_col_unobs,
  s1_col_unobs
)
}
\description{
Process separate dataframes
}
\keyword{internal}
