% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_by_arm_fold}
\alias{split_by_arm_fold}
\title{Split data by arm and fold}
\usage{
split_by_arm_fold(data, n_folds = 2, seed = NULL)
}
\arguments{
\item{data}{An msd_data object}

\item{n_folds}{Number of folds}

\item{seed}{Random seed}
}
\value{
A list with fold assignments for each arm
}
\description{
Split data by arm and fold
}
\keyword{internal}
