% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_design.R
\name{summary.msd_design}
\alias{summary.msd_design}
\title{Summary method for msd_design}
\usage{
\method{summary}{msd_design}(object, ...)
}
\arguments{
\item{object}{An msd_design object.}

\item{...}{Additional arguments passed to the print method.}
}
\value{
Invisibly returns \code{object}.
}
\description{
Summary method for msd_design
}
