% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{validate_and_create}
\alias{validate_and_create}
\title{Validate data and create msd_data object}
\usage{
validate_and_create(observed, unobserved, has_S0, has_S1)
}
\description{
Validate data and create msd_data object
}
\keyword{internal}
