% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{var_pop}
\alias{var_pop}
\title{Compute sample variance (population formula, divide by n)}
\usage{
var_pop(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
Sample variance using n denominator
}
\description{
Compute sample variance (population formula, divide by n)
}
\keyword{internal}
